Stabilize Evo tooling scripts and automation targets
