+++ incoming/lavoro_da_classificare/scripts/README.md
@@ -1,0 +1,5 @@
+# Scripts da classificare
+
+Tutti gli script stabilizzati sono stati spostati in `incoming/scripts/` o
+collegati ai target del `Makefile`. Utilizzare questa directory solo per nuovi
+script in attesa di revisione.
