+++ incoming/scripts/update_tracker_registry.py
